requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-jwt-auth==0.5.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==1.7.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.3.13
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
